# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wlanpi-core",
...
    "version": "2.1.10"
  },
  "paths": {
...
    "/api/v1/system/device/info": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Show Device Info",
        "description": "Returns core information about the PI.\n\nCommands:\n - Uses 'wlanpi-model -b' to query the device model.\n - Uses '/usr/bin/hostname' to query the device hostname.\n - Uses '/etc/wlanpi-release' to query the device software version.\n - Uses '/etc/wlanpi-state' to query the device mode.",
        "operationId": "show_device_info_api_v1_system_device_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/device/stats": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Device Stats",
        "description": "Returns system stats about the PI.\n\nSee get_stats in system_service.py",
        "operationId": "device_stats_api_v1_system_device_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/device/model": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Show Device Model",
        "description": "Uses 'wlanpi-model -b' to query the device model.",
        "operationId": "show_device_model_api_v1_system_device_model_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/service/status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Show A Systemd Service Status",
        "description": "Queries systemd via dbus to get the current status of an allowed service.",
        "operationId": "show_a_systemd_service_status_api_v1_system_service_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/service/start": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Start A Systemd Service",
        "description": "Uses systemd via dbus to start an allowed service.",
        "operationId": "start_a_systemd_service_api_v1_system_service_start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRunning"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/service/stop": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Stop A Systemd Service",
        "description": "Uses systemd via dbus to stop an allowed service.",
        "operationId": "stop_a_systemd_service_api_v1_system_service_stop_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRunning"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/service/restart": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Restart A Systemd Service",
        "description": "Uses systemd via dbus to restart an allowed service.",
        "operationId": "restart_a_systemd_service_api_v1_system_service_restart_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRunning"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/datetime": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Show Datetime",
        "description": "Returns current local date/time and timezone.",
        "operationId": "show_datetime_api_v1_system_datetime_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateTimeInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/timezone": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Show Timezone",
        "description": "Returns the current system timezone.",
        "operationId": "show_timezone_api_v1_system_timezone_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimezoneInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/timezone/list": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "List Timezones",
        "description": "Returns available system timezones.",
        "operationId": "list_timezones_api_v1_system_timezone_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimezoneList"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/timezone/set": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Set Timezone",
        "description": "Sets the system timezone.",
        "operationId": "set_timezone_api_v1_system_timezone_set_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimezoneSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimezoneInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/reg-domain/list": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "List Reg Domains",
        "description": "Returns supported WiFi regulatory domain country codes.",
        "operationId": "list_reg_domains_api_v1_system_reg_domain_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegDomainList"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/reg-domain": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Show Reg Domain",
        "description": "Returns the current WiFi regulatory domain.",
        "operationId": "show_reg_domain_api_v1_system_reg_domain_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegDomainInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/reg-domain/set": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Set Reg Domain",
        "description": "Sets the WiFi regulatory domain country code.",
        "operationId": "set_reg_domain_api_v1_system_reg_domain_set_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegDomainSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegDomainInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/battery": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Show Battery",
        "description": "Returns battery status if a power supply is present.",
        "operationId": "show_battery_api_v1_system_battery_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatteryInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/timezone/auto": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Enable Timezone Auto",
        "description": "Enable NTP automatic time synchronization.",
        "operationId": "enable_timezone_auto_api_v1_system_timezone_auto_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NtpAutoInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/reboot": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Reboot Device",
        "description": "Reboot the device immediately.",
        "operationId": "reboot_device_api_v1_system_reboot_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PowerActionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/shutdown": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Shutdown Device",
        "description": "Shut down the device immediately.",
        "operationId": "shutdown_device_api_v1_system_shutdown_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PowerActionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/hotspot/clients": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Hotspot connected client count",
        "description": "Connected client count for hotspot mode.\n\nReturns 409 when the device is not in hotspot mode.",
        "operationId": "show_hotspot_clients_api_v1_system_hotspot_clients_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "iface",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Iface"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotClients"
                }
              }
            }
          },
          "409": {
            "description": "Device mode precondition not met (e.g. not in hotspot mode)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/hotspot/ssid-passphrase": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Hotspot SSID and WPA passphrase",
        "description": "Hotspot SSID and WPA passphrase from hostapd configuration.\n\nReturns 409 when the device is not in hotspot mode.",
        "operationId": "show_hotspot_ssid_passphrase_api_v1_system_hotspot_ssid_passphrase_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotCredentials"
                }
              }
            }
          },
          "409": {
            "description": "Device mode precondition not met (e.g. not in hotspot mode)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceInfo": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "example": "R4"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "example": "wlanpi-bc2"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname",
            "example": "wlanpi-bc2.local"
          },
          "software_version": {
            "type": "string",
            "title": "Software Version",
            "example": "3.2.0"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "example": "classic"
          }
        },
        "type": "object",
        "required": [
          "model",
          "name",
          "hostname",
          "software_version",
          "mode"
        ],
        "title": "DeviceInfo"
      },
      "DeviceModel": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "example": "R4"
          }
        },
        "type": "object",
        "required": [
          "model"
        ],
        "title": "DeviceModel"
      },
      "DeviceStats": {
        "properties": {
          "ip": {
            "type": "string",
            "title": "Ip",
            "example": "127.0.0.1"
          },
          "cpu": {
            "type": "string",
            "title": "Cpu",
            "example": "23%"
          },
          "ram": {
            "type": "string",
            "title": "Ram",
            "example": "1022/3792MB 26.95%"
          },
          "disk": {
            "type": "string",
            "title": "Disk",
            "example": "6/59GB 11%"
          },
          "cpu_temp": {
            "type": "string",
            "title": "Cpu Temp",
            "example": "1h 40m"
          },
          "uptime": {
            "type": "string",
            "title": "Uptime",
            "example": "1h 40m"
          }
        },
        "type": "object",
        "required": [
          "ip",
          "cpu",
          "ram",
          "disk",
          "cpu_temp",
          "uptime"
        ],
        "title": "DeviceStats"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServiceRunning": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "wlanpi-fpms"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "example": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "active"
        ],
        "title": "ServiceRunning"
      },
      "ServiceStatus": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "wlanpi-fpms"
            ]
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "examples": [
              true
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "active"
        ],
        "title": "ServiceStatus"
      },
...
      "TimezoneInfo": {
        "properties": {
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "example": "Europe/London"
          }
        },
        "type": "object",
        "required": [
          "timezone"
        ],
        "title": "TimezoneInfo"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}